homework1/requirements.txt
numpy
matplotlib
scikit-learn
click
Pillow
pandas
scipy
